moved JSX for edit button into a React component
