+++ src/app/ui/forms/deleteJewelryBtn.tsx
@@ -1,0 +1,8 @@
+
+export default function DeleteJewelryButton() {
+    return (
+        <button className="bg-red-500 hover:bg-red-700 text-white font-bold py-2 px-4 rounded">
+            Delete Jewelry
+        </button>
+    )
+}
